feat(sessions): Add the ingest incremental sums and session update

+++ src/backend/Zeeq.Core.Models/Telemetry/IAgentConversationQueryStore.cs
@@ -114,8 +114,8 @@ public sealed record AgentConversationSummary(
     string? CreatedById,
     DateTimeOffset StartedAtUtc,
     DateTimeOffset? CompletedAtUtc,
-    int TotalInputTokens,
-    int TotalOutputTokens,
+    long TotalInputTokens,
+    long TotalOutputTokens,
     decimal? TotalCostUsd
 );
